requirements-heavy.txt
torch
sentence-transformers
numpy
scikit-learn
google-generativeai